+++ tests/odoo-tests.sh
@@ -1,5 +1,5 @@
 ## Odoo-WooCommerce Sync Tests
-# Last update: 2026-03-10
+# Last update: 2026-08-29
 
 
 # Logs:
@@ -9,19 +9,18 @@
 
 # Settings
 domain="website.com"
-domain_root_path="/home/$domain"
+domain_root_path="/home/${domain}"
 subdomain="erp"
 system_user="website"
-odoo_version="18.0"
 database_name="${system_user}_odoo"
 database_username="${system_user}_odoo_user"
 odoo_addon_name="woocommerce_sync"
 
 # Change current directory
-cd $domain_root_path/domains/$subdomain.$domain/odoo
+cd ${domain_root_path}/domains/${subdomain}.${domain}/odoo
 
 # Delete all queue jobs
-docker exec -i odoo_postgres_${system_user} psql -U $database_username -d $database_name <<EOF
+docker exec -i odoo_postgres_${system_user} psql -U ${database_username} -d ${database_name} <<EOF
 DELETE FROM queue_job;
 EOF
 
@@ -31,8 +30,8 @@ docker restart odoo_server_${system_user}
 # Update WooCommerce Sync module
 docker exec -it odoo_server_${system_user} \
     odoo \
-    --database $database_name \
-    --update $odoo_addon_name \
+    --database ${database_name} \
+    --update ${odoo_addon_name} \
     --no-http \
     --stop-after-init
 
@@ -48,7 +47,7 @@ truncate -s 0 $(docker inspect --format='{{.LogPath}}' odoo_server_${system_user
 # docker logs odoo_postgres_${system_user}
 
 # Access Odoo shell
-# docker exec -it odoo_server_${system_user} odoo shell --no-http -d $database_name
+# docker exec -it odoo_server_${system_user} odoo shell --no-http -d ${database_name}
 
 # Retrieve all field information for product templates
 # fields = self.env['product.template'].fields_get()